+++ aim/web/ui/src/components/MediaPanel/MediaPanel.tsx
@@ -38,6 +38,8 @@ function MediaPanel({
   actionPanel,
   actionPanelSize,
   tooltipType,
+  sortFieldsDict,
+  sortFields,
 }: IMediaPanelProps): React.FunctionComponentElement<React.ReactNode> {
   const [activePointRect, setActivePointRect] = React.useState<{
     top: number;
@@ -159,7 +161,13 @@ function MediaPanel({
   const mediaSetKey = React.useMemo(
     () => Date.now(),
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [data, wrapperOffsetHeight, wrapperOffsetWidth, additionalProperties],
+    [
+      data,
+      wrapperOffsetHeight,
+      wrapperOffsetWidth,
+      additionalProperties,
+      sortFieldsDict,
+    ],
   );
 
   React.useEffect(() => {
@@ -187,7 +195,7 @@ function MediaPanel({
         isLoading={isLoading}
         className='MediaPanel__loader'
         height='100%'
-        loaderComponent={<ChartLoader controlsCount={2} />}
+        loaderComponent={<ChartLoader controlsCount={3} />}
       >
         {panelResizing ? (
           <div className='MediaPanel__Container__resizing'>
@@ -196,32 +204,33 @@ function MediaPanel({
             </Text>
           </div>
         ) : (
-          <>
+          <ErrorBoundary>
             <div className='MediaPanel__Container'>
               {!isEmpty(data) ? (
                 <div
                   className='MediaPanel'
                   style={{ height: `calc(100% - ${actionPanelSize || 0})` }}
                 >
-                  <div
-                    ref={containerRef}
-                    className='MediaPanel__mediaSetContainer'
-                    onMouseOver={onMouseOver}
-                    // TODO
-                    // onClick={(e) => {
-                    //   e.stopPropagation();
-                    //   syncHoverState({
-                    //     activePoint: activePointRef.current,
-                    //     focusedStateActive: false,
-                    //   });
-                    // }}
-                  >
-                    <ErrorBoundary>
+                  <ErrorBoundary>
+                    <div
+                      ref={containerRef}
+                      className='MediaPanel__mediaSetContainer'
+                      onMouseOver={onMouseOver}
+                      // TODO
+                      // onClick={(e) => {
+                      //   e.stopPropagation();
+                      //   syncHoverState({
+                      //     activePoint: activePointRef.current,
+                      //     focusedStateActive: false,
+                      //   });
+                      // }}
+                    >
                       <MediaSet
                         data={data}
                         onListScroll={onListScroll}
                         addUriToList={addUriToList}
                         mediaSetKey={mediaSetKey}
+                        sortFieldsDict={sortFieldsDict}
                         wrapperOffsetHeight={wrapperOffsetHeight - 48}
                         wrapperOffsetWidth={wrapperOffsetWidth}
                         focusedState={focusedState}
@@ -231,26 +240,31 @@ function MediaPanel({
                         tableHeight={tableHeight}
                         tooltip={tooltip}
                         mediaType={mediaType}
+                        sortFields={sortFields}
                       />
-                    </ErrorBoundary>
-                  </div>
-                  {tooltipType && (
-                    <ChartPopover
-                      containerNode={containerRef.current}
-                      activePointRect={activePointRect}
-                      open={
-                        resizeMode !== ResizeModeEnum.MaxHeight &&
-                        !panelResizing &&
-                        (tooltip?.display || focusedState?.active)
-                      }
-                      chartType={tooltipType}
-                      tooltipContent={tooltip?.content}
-                      focusedState={focusedState}
-                    />
-                  )}
-                  {controls && (
-                    <div className='MediaPanel__controls'>{controls}</div>
-                  )}
+                    </div>
+                  </ErrorBoundary>
+                  <ErrorBoundary>
+                    {tooltipType && (
+                      <ChartPopover
+                        containerNode={containerRef.current}
+                        activePointRect={activePointRect}
+                        open={
+                          resizeMode !== ResizeModeEnum.MaxHeight &&
+                          !panelResizing &&
+                          (tooltip?.display || focusedState?.active)
+                        }
+                        chartType={tooltipType}
+                        tooltipContent={tooltip?.content}
+                        focusedState={focusedState}
+                      />
+                    )}
+                  </ErrorBoundary>
+                  <ErrorBoundary>
+                    {controls && (
+                      <div className='MediaPanel__controls'>{controls}</div>
+                    )}
+                  </ErrorBoundary>
                 </div>
               ) : (
                 <EmptyComponent
@@ -260,7 +274,7 @@ function MediaPanel({
               )}
               {actionPanel}
             </div>
-          </>
+          </ErrorBoundary>
         )}
       </BusyLoaderWrapper>
     </ErrorBoundary>